refactor fetch_busco_db.py to database.py
